+++ back/app/Http/Controllers/inventory/RowController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\inventory;
 use Illuminate\Http\Request;
 use App\Helpers\ResponseHelper;
 use App\Http\Controllers\Controller;
+use App\Models\inventory\Product;
 use App\Models\inventory\Row;
 use Illuminate\Pagination\Paginator;
 
@@ -111,6 +112,12 @@ class RowController extends Controller
 
         if (!$data) {
             return ResponseHelper::NoExits('No existe información con el id '.  $id);
+        }
+
+        $product = Product::where('row_id', $id)->first();
+        
+        if ($product) {
+            return ResponseHelper::NoExits('No es posible eliminar la fila ya que está asociada con productos existentes');
         }
 
         $data->delete();
